+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+### [2.2.0] - 2026-05-15
+
+- dep(address-rfc2822): replaced by @haraka/email-address
+- test: runner is now node:test
+- test: convert done callbacks to async/await #20
+
 ### [2.1.2] - 2026-03-03
 
 - fix: isa was boolean, revert change from #4 to yes/no
@@ -137,3 +143,4 @@ Changes to README.md
 [2.1.1]: https://github.com/haraka/haraka-plugin-bounce/releases/tag/v2.1.1
 [2.1.0]: https://github.com/haraka/haraka-plugin-bounce/releases/tag/v2.1.0
 [2.1.2]: https://github.com/haraka/haraka-plugin-bounce/releases/tag/v2.1.2
+[2.2.0]: https://github.com/haraka/haraka-plugin-bounce/releases/tag/v2.2.0